backend/pyproject.toml
[tool.poetry]
name = "whatdo2"
version = "0.1.0"
description = ""
authors = ["Mikey Mo <mikeyjkmo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.20.0"
dnspython = "^2.2.1"
pydantic = "^1.9.0"
fastapi = "^0.75.2"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
motor = "^3.0.0"
SQLAlchemy = {extras = ["asyncio", "mypy"], version = "^1.4.36"}
psycopg2 = "^2.9.3"
asyncpg = "^0.25.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pytest-watch = "^4.2.0"
pytest-icdiff = "^0.5"
black = "^22.3.0"
flake8 = "^4.0.1"
mypy = "^0.950"
pytest-asyncio = "^0.18.3"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
